Added service and test to delete role by id

+++ src/modules/roles/services/delete-role/delete-role-by-id.use-case.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { Result } from "src/common/result/result";
 import { IRoleRepository } from "../../repository/iroles.repository";
+import { isUUID } from "class-validator";
 
 @Injectable()
 export class DeleteRoleByIdUseCase {
@@ -9,6 +10,7 @@ export class DeleteRoleByIdUseCase {
     ) {}
 
     async execute(id: string): Promise<Result<void>> {
+        if (!isUUID(id)) return Result.badRequest('Id should be a UUID');
 
         const result = await this.roleRepository.deleteById(id)
         if (result === false) {
